+++ library/src/main/java/nl/remcoder/adventofcode/library/pathfinding/Node.java
@@ -3,15 +3,15 @@ package nl.remcoder.adventofcode.library.pathfinding;
 import java.util.Map;
 
 public interface Node extends Comparable<Node> {
-    Map<? extends Node, Integer> getNeighbors();
-    int getDistance();
+    Map<? extends Node, Long> getNeighbors();
+    long getDistance();
     boolean isVisited();
     void setVisited(boolean visited);
-    void setDistance(int distance);
+    void setDistance(long distance);
     void printStateInformation();
 
     @Override
     default int compareTo(Node o) {
-        return Integer.compare(this.getDistance(), o.getDistance());
+        return Long.compare(this.getDistance(), o.getDistance());
     }
 }
